+++ CHANGELOG.md
@@ -41,7 +41,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * `config`:
   * Added `webpack.mode`
-  * Added `webpack.cache`
+  * Added `webpack.cacheType`
 
 * Added API to load the dynamic dependencies `iBlock`
 * Added `StaticRouteMeta.load` `core/router`