+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.102 (2020-11-27)
+
+#### :bug: Bug Fix
+
+* Fixed layout shifts after `reInit` `base/b-virtual-scroll`
+
 ## v3.0.0-rc.101 (2020-11-18)
 
 #### :bug: Bug Fix